% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascript.R
\docType{data}
\name{myData}
\alias{myData}
\title{A fictional dataset for air pollution in multiple cities}
\format{
A data frame with 41 rows and 7 variables:
\describe{
\item{y}{Overall pollution level in the city}
\item{x1}{dangerous air pollutant no. 1}
\item{x2}{dangerous air pollutant no. 2}
\item{x3}{particular matter in air}
\item{x4}{very dangerous pollutant}
\item{x5}{dangerous pollutant found in river water}
\item{x6}{dangerous pollutant found in soil}
...
}
}
\source{
\url{http://canvas.vt.edu/}
}
\usage{
myData
}
\description{
The dataset simulates levels of specific pollutants measured in several world
capitals on the same data May 20, 2015.
}
\keyword{datasets}
