% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Run regression of Y on X}
\usage{
myLinearRegression(X = NULL, Y = NULL, sub = c(1:length(Y)))
}
\arguments{
\item{X}{A matrix of covariates}

\item{Y}{A vector of outcomes}

\item{sub}{A list of subjects (i.e. a set of integers corresponding to rows in X)}
}
\value{
A list containing the coefficients form the linear regression of Y on X and corresponding p-values.
}
\description{
This will run a linear regression of Y on X and produce a scattergram, coefficients & corresponding p-values.
}
