% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myData}
\alias{myData}
\title{Edgar Anderson's Iris Data}
\format{
A data frame with 150 rows and 5 variables:
\describe{
\item{ID}{this id}
\item{X}{this x}
...
}
}
\source{
iris datasets from base R

Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179–188.
}
\usage{
myData
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables
sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris.
The species are Iris setosa, versicolor, and virginica.
}
\keyword{datasets}
