% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myData}
\alias{myData}
\title{Sample Data for Linear Regression Analysis}
\format{
A data frame with 100 rows and 5 variables:
\describe{
\item{Y}{Correlation of all X columns}
\item{X}{Temperature}
\item{X}{Amount of rainfall each day}
\item{X}{Amount of snowfall each day}
\item{X}{Humidity}
\item{X}{UV Index}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
myData
}
\description{
This data is meant to be used as an example for inputs to myLinearRegression. Written by Sara Khorramshahgol
}
\keyword{datasets}
