% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myData.R
\docType{data}
\name{myData}
\alias{myData}
\title{myData Datasey}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{WaterLevel}{Numeric. River gauge level measured in feet of
difference from average of last 365 days}
\item{Rainfall}{Numeric. Measured in inches difference for last 10 days
from average for last 365 days.}
\item{Turbidity}{Numeric. Measured in light transmissivity difference over last 10 days difference
from average for last 365 days.}
\item{TempSurface}{Numeric. Measured in degrees F difference over last 10 days from average for last 365 days.}
\item{TempBottom}{Numeric. Measured in degrees F difference over last 10 days from average for last 365 days.}
\item{Sediment}{Numeric. Micrograms/cubic cm. Measured as difference over last 10 days from average for last 365 days.}
}
\usage{
myData
}
\description{
This is an artificially engineered data set you can use for
testing the myLinearRegression function in the
myRegressionPackage.
}
\details{
#'
}
\keyword{datasets}
