% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Creates a scatterplot of each pair of covariates.
Runs a linear regression of Y on X.}
\usage{
myLinearRegression(X, Y, sub = 0)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{Y}{A vector of outcomes.}

\item{sub}{a list of subjects (i.e. a set of integers corresponding to rows in X)}
}
\value{
coef: a set of coefficients from the linear regression of Y on X

pvals: the set of corrresponding p-values
Imports: ggplot2
}
\description{
Created by Emilios Chrysostomides
}
\examples{
myData<- read.csv("myData.csv", header = TRUE)
Y <- myData[,1]
X <- data.matrix(myData[,-1], rownames.force = NA)
myLinearRegression(X, Y)
myLinearRegression(X[,-1], Y)
myLinearRegression(X[,-1], Y, c(1:5))
}
