% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Linear Regression Analysis}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{X}{A Matrix of covariates.}

\item{sub}{A List of subjects.}
}
\value{
A list contains coefficients and p-values.
}
\description{
This is a function that run linear regression analysis on the Y on X, and paired graphs of variables.
}
\examples{
myLinearRegression(myData[,1],myData[,2:5],1:10)
myLinearRegression(myData[,1],myData[,2:3],1:20)
}
