% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Linear Regression Analysis}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{X:}{a matrix of  covariates}

\item{Y:}{a vector  of outcomes}

\item{sub:}{a list of  subjects (i.e.  a set of integers corresponding to rows in X)}
}
\value{
coef: a set  of coefficients  from  the linear  regression of  Y on X

pvals: the set of corrresponding p-values
}
\description{
Creates  a scatterplot  of  each pair  of covariates. Also creates a linear regression of Y on X. Written by Sara Khorramshahgol
}
