% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{My Linear Regression}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{A vector of outcomes}

\item{X}{A matrix of covariates}

\item{sub}{a list of  subjects (i.e.  a set of integers corresponding to rows in X)}
}
\value{
Returns the coefficients of the linear regression of X and Y and the corresponding p-values
}
\description{
My Linear Regression
}
\examples{


}
