% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{create a linear regression model}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{a vector of outcomes}

\item{X}{a matrix of covariates}

\item{sub}{indices of the rows to be used in the analisys}
}
\value{
a list with two vectors:
coef - coefficients from the linear regression formula
pvals - p-values corresponding to the coefficients
}
\description{
This function returns the parameters of the linear regression model for the given data
}
\examples{
myLinearRegression(data$outcome, data[1:4], 1:100)
}
