% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Creates plot and runs linear regression}
\usage{
myLinearRegression(x, y, sub)
}
\arguments{
\item{x}{A matrix of covariates.}

\item{y}{A vector of outcomes}

\item{sub}{A list of subjects}
}
\value{
scatterplot of \code{x} and linear regression \code{y} an \code{x}.
}
\description{
This function creates a scatterplot if the number of subjects is less than 5 and not
greater than 5. If there are 5 or less variables, a linear regression will be performed
on the list of subjects specified, or sub.The linear regression will output a list of the
coefficients and the p-values.
}
\examples{
aGreatFunction(x1, y, sub)
}
