% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Create a linear model between y and x(s). This also creates a scatterplot of the y and x variables.}
\usage{
myLinearRegression(x = NULL, y = NULL, sub = NULL)
}
\arguments{
\item{x}{A vector or matrix of data to predict y}

\item{y}{A vector of numbers}

\item{sub}{The number of rows in the data}
}
\value{
Scattergram of variables. Coeffecients and p-values of linear model.
}
\description{
This is truly a great and much-needed function
}
\examples{
myLinearRegression(x, y)
}
