% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Perform Linear Regression on user provided dependent variable Y and
covariates X, i.e. Y = B0+B1x1+...+BnXn}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{A vector of outcomes. Numeric.}

\item{X}{A matrix of covariates. Numeric.
Columns correspond to covariates and rows correspond to
values. Maximum 5 covariates allowed.}

\item{sub}{a set of integers corresponding to rows in X.
Only the specified rows will be used for modeling and
plotting.}
}
\value{
A list with three elements:
\enumerate{
\item plot: A plot object containing the pairwise scatter plots
or a character string indicating there are too many covariates to
plot.
\item coefficients: A vector of model coefficients
\item pvals: A vector of p-values for each coeffcient in the model
}
}
\description{
If the number of covariates is 5 or fewer, scatter plots of
Y input and each pair of covariates will be produced. Otherwise the
function will return a character string "Too may variables to plot. Maximum is 5."
}
\examples{
#Compute a 2-covariate linear regression y~x1+x2, for rows 1-100
#inclusive in y, x1, x2, and then print the pairwise scatter plot 
#returned by the function.
#  
index<-1:100
myData<-data.frame(cbind(y,x1,x2)
model<-myLinearRegression(myData$y,myData[,2:3],index)
model$plot
#
#
#Compute a 1-covariate linear regression y~x1, for rows with selected 
#color category and then print the coefficients returned.
#  
colors<-rep(c("Red","Blue","Green","Yellow"),25)
myData<-data.frame(cbind(y,x1,colors))
index<-index<-as.numeric(rownames(subset(myData,color=="Red")))
xout<-myLinearRegression(myData$y,myData[,2:3],index)
xout$coefficients
}
